+++ 01.workspace/heave/src/fun/sqlite_map_row_to_attribute.rs
@@ -4,7 +4,7 @@ pub fn run(row: &rusqlite::Row) -> rusqlite::Result<Attribute> {
     let id: String = row.get(0)?;
     let _entity_id: String = row.get(1)?;
     let signed_int: Option<i64> = row.get(2)?;
-    let unsigned_int: Option<u64> = row.get(3)?;
+    let unsigned_int: Option<u32> = row.get(3)?;
     let real: Option<f64> = row.get(4)?;
     let text: Option<String> = row.get(5)?;
     let bool: Option<bool> = row.get(6)?;
