+++ 01.workspace/heave/src/fun/sqlite_map_row_to_entity.rs
@@ -4,7 +4,7 @@ pub fn run(row: &rusqlite::Row) -> rusqlite::Result<Entity> {
     let id: String = row.get(0)?;
     let class: String = row.get(1)?;
     let subclass: Option<String> = row.get(2)?;
-    let ref_date: Option<u64> = row.get(3)?;
+    let ref_date: Option<u32> = row.get(3)?;
     let entity = Entity {
         id,
         state: EntityState::Loaded,
